fixing incorrect release behavior
