+++ producer/build.sbt
@@ -1,6 +1,6 @@
 import sbt._, Keys._
-import dependencies._
 import pulse.plugin._
+import pulse.plugin.dependencies._
 import local.dependencies._
 
 libraryDependencies ++= Seq (
